core/serializer/user_serializer.py
from django.db import models
from rest_framework import serializers
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from django.contrib.auth import get_user_model

from core.models import Code

User = get_user_model()

class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):

    def validate(self, attrs):
        data = super().validate(attrs)
        data['id'] = self.user.id
        data['phone'] = self.user.phone
        data['email'] = self.user.email
        data['is_staff'] = self.user.is_staff
        data['full_name'] = self.user.full_name
        return data

class SignUpSerializer (serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"
        extra_kwargs = {
            'password': {'write_only': True},
        }